+++ dev/babel.config.js
